+++ src/haven_types.h
@@ -4,12 +4,22 @@
 #include <Rcpp.h>
 #include <cmath>
 
-enum FileType {
+enum FileVendor {
   HAVEN_SPSS,
   HAVEN_STATA,
-  HAVEN_SAS,
+  HAVEN_SAS
+};
+
+enum FileExt {
+  HAVEN_SAV,
+  HAVEN_POR,
+  HAVEN_DTA,
+  HAVEN_SAS7BDAT,
+  HAVEN_SAS7BCAT,
   HAVEN_XPT
 };
+
+FileVendor extVendor(FileExt ext);
 
 enum VarType {
   HAVEN_DEFAULT,
@@ -18,15 +28,15 @@ enum VarType {
   HAVEN_DATETIME
 };
 
-std::string formatAttribute(FileType type);
+std::string formatAttribute(FileVendor vendor);
 bool hasPrefix(std::string x, std::string prefix);
 VarType numType(SEXP x);
-VarType numType(FileType type, const char* var_format);
+VarType numType(FileVendor vendor, const char* var_format);
 
 // Value conversion -----------------------------------------------------------
 
-int daysOffset(FileType type);
-double adjustDatetimeToR(FileType file, VarType var, double value);
-double adjustDatetimeFromR(FileType file, SEXP col, double value);
+int daysOffset(FileVendor vendor);
+double adjustDatetimeToR(FileVendor vendor, VarType var, double value);
+double adjustDatetimeFromR(FileVendor vendor, SEXP col, double value);
 
 #endif